+++ hydrodiy/gis/c_grid.h
@@ -6,23 +6,23 @@
 #define GRID_ERROR 50000
 
 double clipd(double x, double x0, double x1);
-long clipi(long x, long x0, long x1);
+long long clipi(long long x, long long x0, long long x1);
 
-long getnxy(long ncols, long idxcell, long *nxy);
+long long getnxy(long long ncols, long long idxcell, long long *nxy);
 
-long c_coord2cell(long nrows, long ncols,
+long long c_coord2cell(long long nrows, long long ncols,
     double xll, double yll, double csz,
-    long nval, double * xycoords, long * idxcell);
+    long long nval, double * xycoords, long long * idxcell);
 
 
-long c_cell2coord(long nrows, long ncols,
+long long c_cell2coord(long long nrows, long long ncols,
     double xll, double yll, double csz,
-    long nval, long * idxcell, double * xycoords);
+    long long nval, long long * idxcell, double * xycoords);
 
 
-long c_slice(long nrows, long ncols,
+long long c_slice(long long nrows, long long ncols,
     double xll, double yll, double csz, double* data,
-    long nval, double* xyslice, double * zslice);
+    long long nval, double* xyslice, double * zslice);
 
-long c_neighbours(long nrows, long ncols,
-    long idxcell, long * neighbours);
+long long c_neighbours(long long nrows, long long ncols,
+    long long idxcell, long long * neighbours);
